+++ src/gas_metering/mod.rs
@@ -36,6 +36,9 @@ pub trait Rules {
 	/// code into the function calling `memory.grow`. Therefore returning anything but
 	/// [`MemoryGrowCost::Free`] introduces some overhead to the `memory.grow` instruction.
 	fn memory_grow_cost(&self) -> MemoryGrowCost;
+
+	/// A surcharge cost to calling a function that is added per local of that function.
+	fn call_per_local_cost(&self) -> u32;
 }
 
 /// Dynamic costs for memory growth.
@@ -76,6 +79,7 @@ impl MemoryGrowCost {
 pub struct ConstantCostRules {
 	instruction_cost: u32,
 	memory_grow_cost: u32,
+	call_per_local_cost: u32,
 }
 
 impl ConstantCostRules {
@@ -83,15 +87,15 @@ impl ConstantCostRules {
 	///
 	/// Uses `instruction_cost` for every instruction and `memory_grow_cost` to dynamically
 	/// meter the memory growth instruction.
-	pub fn new(instruction_cost: u32, memory_grow_cost: u32) -> Self {
-		Self { instruction_cost, memory_grow_cost }
+	pub fn new(instruction_cost: u32, memory_grow_cost: u32, call_per_local_cost: u32) -> Self {
+		Self { instruction_cost, memory_grow_cost, call_per_local_cost }
 	}
 }
 
 impl Default for ConstantCostRules {
 	/// Uses instruction cost of `1` and disables memory growth instrumentation.
 	fn default() -> Self {
-		Self { instruction_cost: 1, memory_grow_cost: 0 }
+		Self { instruction_cost: 1, memory_grow_cost: 0, call_per_local_cost: 1 }
 	}
 }
 
@@ -102,6 +106,10 @@ impl Rules for ConstantCostRules {
 
 	fn memory_grow_cost(&self) -> MemoryGrowCost {
 		NonZeroU32::new(self.memory_grow_cost).map_or(MemoryGrowCost::Free, MemoryGrowCost::Linear)
+	}
+
+	fn call_per_local_cost(&self) -> u32 {
+		self.call_per_local_cost
 	}
 }
 
@@ -249,8 +257,16 @@ pub fn inject<R: Rules, B: Backend>(
 							}
 						}
 					}
-					if inject_counter(func_body.code_mut(), gas_fn_cost, rules, gas_func_idx)
-						.is_err()
+					let locals_count =
+						func_body.locals().iter().map(|val_type| val_type.count()).sum();
+					if inject_counter(
+						func_body.code_mut(),
+						gas_fn_cost,
+						locals_count,
+						rules,
+						gas_func_idx,
+					)
+					.is_err()
 					{
 						error = true;
 						break
@@ -566,6 +582,7 @@ fn add_grow_counter<R: Rules>(
 fn determine_metered_blocks<R: Rules>(
 	instructions: &elements::Instructions,
 	rules: &R,
+	locals_count: u32,
 ) -> Result<Vec<MeteredBlock>, ()> {
 	use parity_wasm::elements::Instruction::*;
 
@@ -573,6 +590,9 @@ fn determine_metered_blocks<R: Rules>(
 
 	// Begin an implicit function (i.e. `func...end`) block.
 	counter.begin_control_block(0, false);
+	// Add locals initialization cost to the function block.
+	let locals_init_cost = rules.call_per_local_cost().checked_mul(locals_count).ok_or(())?;
+	counter.increment(locals_init_cost)?;
 
 	for cursor in 0..instructions.elements().len() {
 		let instruction = &instructions.elements()[cursor];
@@ -640,10 +660,11 @@ fn determine_metered_blocks<R: Rules>(
 fn inject_counter<R: Rules>(
 	instructions: &mut elements::Instructions,
 	gas_function_cost: u64,
+	locals_count: u32,
 	rules: &R,
 	gas_func: u32,
 ) -> Result<(), ()> {
-	let blocks = determine_metered_blocks(instructions, rules)?;
+	let blocks = determine_metered_blocks(instructions, rules, locals_count)?;
 	insert_metering_calls(instructions, gas_function_cost, blocks, gas_func)
 }
 
@@ -668,7 +689,8 @@ fn insert_metering_calls(
 		// If there the next block starts at this position, inject metering instructions.
 		let used_block = if let Some(block) = block_iter.peek() {
 			if block.start_pos == original_pos {
-				new_instrs.push(I64Const((block.cost + gas_function_cost) as i64));
+				new_instrs
+					.push(I64Const((block.cost.checked_add(gas_function_cost).ok_or(())?) as i64));
 				new_instrs.push(Call(gas_func));
 				true
 			} else {
@@ -722,7 +744,7 @@ mod tests {
 		);
 		let backend = host_function::Injector::new("env", "gas");
 		let injected_module =
-			super::inject(module, backend, &ConstantCostRules::new(1, 10_000)).unwrap();
+			super::inject(module, backend, &ConstantCostRules::new(1, 10_000, 1)).unwrap();
 
 		assert_eq!(
 			get_function_body(&injected_module, 0).unwrap(),
@@ -759,7 +781,7 @@ mod tests {
 		);
 		let backend = mutable_global::Injector::new("gas_left");
 		let injected_module =
-			super::inject(module, backend, &ConstantCostRules::new(1, 10_000)).unwrap();
+			super::inject(module, backend, &ConstantCostRules::new(1, 10_000, 1)).unwrap();
 
 		assert_eq!(
 			get_function_body(&injected_module, 0).unwrap(),